+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- a local `k3d` development workflow for creating and managing CES test environments with local k3d-clusters.
 
 ## [v6.0.0] - 2026-04-13
 ### Changed